Add back strong name signing
When we made the transition to use dotnet cli to build the project we
stopped strong name signing the resulting binaries.  This caused a
problem when we try and use these binaries to build our perf tests in
the CoreFX repo.  To fix this problem we just use the
GlobablAssemblyInfo file to apply the key to all of the projects.  We
also neede to update the AssemblyInfo of the execution project because
it allows the api project to view its internal state and since the
binaries are now strong name signed we need to include the public key.