Adding a Ctrl-C, Ctrl-Break, Close handle for disposable objects.
The Problem:
When the running harness was stopped with a console control signal such as: Ctrl-C, Ctrl-Break, or Close signals, the trace event session did not clean up properly, the handles to the etl files and sessions themselves persisted.

The Solution:
Wraps the trace event sessions in an object that subscribe them to a Ctrl listener, and dispose them accordingly when the signal is received.

Note:
Killing the process under the debugger, with the signals specified above, does not make it very happy. The process seems continue running or crashes.