Add ability to run larger benchmarks
We add new functionality to the xunitPerformanceApi to support custom
benchmarks. The function RunStartupBenchmark receives arguments an three
delegates Pre, Mid and Post to be called before all the runs, after each
one and after finishing.
Via this delegates, we let the test authors to produce and collect the
measuring for its runs however they want.

We add a new class PerformanceTestConfig to collect all the
configuration needed for launching this tests. It's been created
thinking in JitBench, but it can be modified to include more options and
setups.